chore(components/forms + services)[annoucenemnt-management-form + sequential-turns-management-form + announcement-group-subgroup-management]: rename the directory of service
