+++ src/app/components/forms/announcement-management-form/announcement-group-form/announcement-group-form.component.ts
@@ -17,8 +17,8 @@ import { ValidationSchema } from 'app/constants/validation-schema';
 import { checkAndToastError } from 'app/utils/api-utils';
 import { LoadingService } from 'app/services/loading-service/loading-service.service';
 import { ToastService } from 'app/services/toast-service/toast.service';
-import { AnnouncementGroupSubgroupManagementService } from 'app/services/announcement_group_subgroup_management/announcement-group-subgroup-management.service';
-import { AnnouncementGroup } from 'app/services/announcement_group_subgroup_management/model/announcement-group.model';
+import { AnnouncementGroupSubgroupManagementService } from 'app/services/announcement-group-subgroup-management/announcement-group-subgroup-management.service';
+import { AnnouncementGroup } from 'app/services/announcement-group-subgroup-management/model/announcement-group.model';
 import { ButtonComponent } from 'app/components/shared/button/button.component';
 import { TableConfig } from 'app/constants/ui/table.ui';
