+++ src/app/components/forms/sequential-turns-management-form/relation-of-sequential-turn-to-announcement-sub-groups/relation-of-sequential-turn-to-announcement-sub-groups-form.component.ts
@@ -7,9 +7,9 @@ import {
 } from '@angular/forms';
 import { SearchAutoCompleteComponent } from 'app/components/shared/inputs/search-auto-complete/search-auto-complete.component';
 import { TableConfig } from 'app/constants/ui/table.ui';
-import { AnnouncementGroupSubgroupManagementService } from 'app/services/announcement_group_subgroup_management/announcement-group-subgroup-management.service';
-import { AnnouncementGroup } from 'app/services/announcement_group_subgroup_management/model/announcement-group.model';
-import { AnnouncementSubGroup } from 'app/services/announcement_group_subgroup_management/model/announcement-subgroup.model';
+import { AnnouncementGroupSubgroupManagementService } from 'app/services/announcement-group-subgroup-management/announcement-group-subgroup-management.service';
+import { AnnouncementGroup } from 'app/services/announcement-group-subgroup-management/model/announcement-group.model';
+import { AnnouncementSubGroup } from 'app/services/announcement-group-subgroup-management/model/announcement-subgroup.model';
 import { LoadingService } from 'app/services/loading-service/loading-service.service';
 import { RelationOfSequentialTurnToAnnouncementSubGroup } from 'app/services/sequential-turn-management/model/relation-of-sequentialTurn-to-announcementSubGroup.model';
 import { SequentialTurn } from 'app/services/sequential-turn-management/model/sequential-turn.model';
