+++ src/app/services/announcement-group-subgroup-management/model/announcement-group.model.ts
