+++ src/app/services/announcement-group-subgroup-management/model/relation-of-announcement-group-subgroup.model.ts
